Remove "mode" from followup
Removed mode from followup asset content. It was logically impossible to have a mode with a metric. Falling back to just asset type.
