+++ apps/web/src/layouts/ChatLayout/FileContainer/FileContainerHeader/DashboardContainerHeaderButtons/DashboardContainerHeaderButtons.tsx
@@ -84,7 +84,7 @@ AddContentToDashboardButton.displayName = 'AddContentToDashboardButton';
 
 const FollowUpWithAssetButton = React.memo(({ dashboardId }: { dashboardId: string }) => {
   return (
-    <FollowUpWithAssetPopup assetId={dashboardId} assetType="dashboard" mode="filter">
+    <FollowUpWithAssetPopup assetId={dashboardId} assetType="dashboard">
       <Button variant="ghost" prefix={<BarsFilter />} />
     </FollowUpWithAssetPopup>
   );
